Bug 1729170 - Make all methods of `nsTextFragment` take `uint32_t` for the offset and length of text r=smaug
`nsTextFragment` is a storage of data node and DOM offset is `uint32_t`, but
some methods of `nsTextFragment` takes `int32_t` for the offset/length in
its text.  Therefore, callers need to cast from `uint32_t` to `int32_t` if
the offset value is offset in a DOM node.  Therefore, all methods of it should
take `uint32_t` values as offset/length in its text.

Differential Revision: https://phabricator.services.mozilla.com/D127923

UltraBlame original commit: 05d338fecf52e15beb4e30f113910cb50f648cb8

+++ toolkit/components/typeaheadfind/nsTypeAheadFind.cpp
@@ -840,10 +840,8 @@ void nsTypeAheadFind::RangeStartsInsideLink(nsRange* aRange,
       
       for (uint32_t index = 0; index < startOffset; index++) {
         
-        if (!mozilla::dom::IsSpaceCharacter(
-                textFrag->CharAt(static_cast<int32_t>(index)))) {
+        if (!mozilla::dom::IsSpaceCharacter(textFrag->CharAt(index))) {
           *aIsStartingLink = false;  
-
           break;
         }
       }
